Adapt to damaged rects (#5)
* Expose damagedRects as a member of PerformanceContainerTiming events

As native implementation exposes damagedRects in PerformanceContainerTiming
for obtaining the set of rectangles representing the damaged area, do the
same in the polyfill. As we already calculated that information in paintedRects
convert the old implementation to match the new API.

* Ignore paintDebugOverlay generated divs in ContainerTiming

Add the containertiming-ignore property to the divs for the
debug overlay so they are not accounted in any of the container
timing entries, even if it is the toplevel html or body.

* Support native case paint highlight

To support painting the damaged areas in the native case, call
the method directly from the test cases, and remove automatic
ctDebug for those cases.