Document the initial proposal for containertiming-nesting
Though the explainer already talks about the problem of defining
different policies for nesting of the container timing events in
nested container roots, there is not proposal for how to adjust
that.

The initial Chromium implementation patches propose the new
attribute containertiming-nesting for that purpose. Add to the
explainer how this initial implementation works.