Updates for ContainerTiming nesting and recent Chromium patches (#23)
* Allow to show more than one performance observer entry

In the demos and the extension, after moving to use the demo-overlays
we were only showing the last of the performance entries.

* Remove nesting strategy PerformanceObserver init parameter

* Add HTML attribute based nested strategy selection

* Document the initial proposal for containertiming-nesting

Though the explainer already talks about the problem of defining
different policies for nesting of the container timing events in
nested container roots, there is not proposal for how to adjust
that.

The initial Chromium implementation patches propose the new
attribute containertiming-nesting for that purpose. Add to the
explainer how this initial implementation works.

* Update the implementation work patches list

Most of the initial implementation has landed to Chromium. Refer
to the final patches list.

* Fix the link for the Chromium extension

Chromium extension is now part of this same repository. Update the
link accordingly.